backend/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-CKEditor==0.5.1
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
pymongo==4.7.1
python-dotenv==1.0.1
Werkzeug==3.0.3
minio==7.1.15
PyJWT==2.10.1
Flask-Mail==0.10.0